fix mac compilation

+++ core/iwasm/libraries/thread-mgr/thread_manager.c
@@ -599,11 +599,11 @@ thread_manager_start_routine(void *arg)
 
     os_mutex_lock(&exec_env->wait_lock);
 #if WASM_ENABLE_CHECKPOINT_RESTORE != 0
-    uint64_t old_handle = exec_env->handle;
+    uint64_t old_handle = (uint64_t)exec_env->handle;
 #endif
     exec_env->handle = os_self_thread();
 #if WASM_ENABLE_CHECKPOINT_RESTORE != 0
-    wamr_handle_map(old_handle, exec_env->handle);
+    wamr_handle_map(old_handle, (uint64_t)exec_env->handle);
     //wamr_korp_tid_map(old_korp_tid, os_self_thread());
 #endif
     /* Notify the parent thread to continue running */
